Refactor OptionDLCType
