+++ core/src/main/scala/org/bitcoins/core/protocol/tlv/DLCSpecTypeSerializable.scala
@@ -1,0 +1,10 @@
+package org.bitcoins.core.protocol.tlv
+
+import org.bitcoins.crypto.NetworkElement
+import scodec.bits.ByteVector
+
+trait DLCSpecTypeSerializable[+T <: DLCSpecType] extends NetworkElement {
+  def toSubType: T
+
+  override def bytes: ByteVector = toSubType.bytes
+}
