+++ core/src/main/scala/org/bitcoins/core/protocol/tlv/FactoryOptionTLV.scala
@@ -1,0 +1,9 @@
+package org.bitcoins.core.protocol.tlv
+
+import org.bitcoins.crypto.{Factory, NetworkElement}
+import scodec.bits.ByteVector
+
+trait FactoryOptionTLV[T <: NetworkElement] extends Factory[OptionDLCType[T]] {
+
+  override def fromBytes(bytes: ByteVector): OptionDLCType[T]
+}
