+++ testkit/src/main/scala/org/bitcoins/testkit/wallet/DLCWalletUtil.scala
@@ -293,7 +293,7 @@ object DLCWalletUtil extends Logging {
   lazy val sampleContractDataDb: DLCContractDataDb = DLCContractDataDb(
     dlcId = sampleDLCDb.dlcId,
     oracleThreshold = 1,
-    oracleParamsTLVOpt = NoneTLV,
+    oracleParamsTLVOpt = NoneDLCType,
     contractDescriptorTLV = sampleContractDescriptor.toSubType,
     contractMaturity = BlockTimeStamp(0),
     contractTimeout = BlockTimeStamp(1),
